Delete README.md
Why was this added when I pushed two files, the program and readme for Week 1 Exercise 2? I have no idea. It doesn't need to be here. It also says added 2 hours ago. It was committed to Wk1Ex1 two hours ago, and this repository didn't exist then. Will deleting it here delete it in Wk1Ex1? I sure hope not!